[ALPS07779043] media: fix build error
Type error introduced by new clang

MTK-Commit-Id: 90ad744b5ce8536fa8bca6e2804c325a63b9c883

CR-Id: ALPS07779043
Feature: Kernel SI Operation

+++ drivers/media/platform/mtk-isp/camsys/isp7sp/unit_test/mtk_cam_ut-seninf.c
@@ -176,7 +176,7 @@ static int mtk_ut_seninf_of_probe(struct platform_device *pdev,
 		dev_info(dev, "failed to map register base\n");
 		return PTR_ERR(seninf->base);
 	}
-	dev_info(dev, "seninf, map_addr=0x%lx\n", seninf->base);
+	dev_info(dev, "seninf, map_addr=0x%lx\n", (unsigned long)seninf->base);
 
 	clks = of_count_phandle_with_args(pdev->dev.of_node,
 				"clocks", "#clock-cells");
